% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{binMultiNMR}
\alias{binMultiNMR}
\title{A function for binning multiple NMR spectra.}
\usage{
binMultiNMR()
}
\value{
{None}
}
\description{
This function creates bins for each spectrum in mrbin.env$mrbinparam$NMRfolders and
saves the bins to mrbin.env$bins.
}
\examples{
setParam(parameters=list(dimension="1D",binwidth1D=.05,PCA="No",
         NMRfolders=c(system.file("extdata/1/10/pdata/10",package="mrbin"),
                      system.file("extdata/2/10/pdata/10",package="mrbin"))))
binMultiNMR()
}
