% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{binSingleNMR}
\alias{binSingleNMR}
\title{A function for binning a single NMR spectrum.}
\usage{
binSingleNMR()
}
\value{
{None}
}
\description{
This function creates bins for the current spectrum. This function is
meant only for use within the mrbin function.
}
\examples{
\donttest{ binSingleNMR() }
}
