% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{calculateNoise}
\alias{calculateNoise}
\title{A function for calculating noise levels.}
\usage{
calculateNoise()
}
\description{
This function calculates noise levels.
}
\examples{
## Requires data
# calculateNoise()
}
