% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{checkmrbin}
\alias{checkmrbin}
\title{A function for checking mrbin objects.}
\usage{
checkmrbin(mrbinObject, verbose = TRUE, errorsAsWarnings = NULL)
}
\arguments{
\item{mrbinObject}{An mrbin object}

\item{verbose}{Should a summary be displayed? (Warnings will be displayed even when setting verbose to FALSE)}

\item{errorsAsWarnings}{If TRUE, errors will be turned into warnings. Should be used with care, as errors indicate undocumented changes to the data. If not provided, this will be taken from the mrbinObject.}
}
\value{
An (invisible) character vector of warnings
}
\description{
This function checks an mrbin object and returns warning if changes were not documented
}
\examples{
mrbinObject<-createmrbin()
mrbinObject<-checkmrbin(mrbinObject)
}
