% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{contMin}
\alias{contMin}
\title{A function for changing plotNMR plots.}
\usage{
contMin()
}
\value{
{None}
}
\description{
This function decreases the minimum contour level of the current 2D NMR
spectrum plot.
}
\examples{
setParam(parameters=list(dimension="2D",binwidth2D=1,binheight=10,
         NMRfolders=system.file("extdata/1/12/pdata/10",package="mrbin")))
binMultiNMR()
contMin()
}
