% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{createAnnotationList}
\alias{createAnnotationList}
\title{A function for creating annotation lists.}
\usage{
createAnnotationList(
  metaboliteNames,
  binwidth = 0.01,
  binheight = 1,
  verbose = TRUE
)
}
\arguments{
\item{metaboliteNames}{Vector of metabolite names}

\item{binwidth}{Full width of each bin. Will only be used if mrbinObject is a matrix}

\item{binheight}{Full height of each bin. Will only be used if mrbinObject is a matrix and dimension is set to "2D"}

\item{verbose}{Should outputs be shown or suppressed?}
}
\value{
A numeric matrix, the first columns containing metabolite names and the first row being a header. Each row belongs to one unique metabolite signal (left, right, top, bottom borders). Row names are metabolite names.
}
\description{
This function creates a matrix of peak positions for use with the annotatemrbin function
}
\examples{
resetEnv()#clean up previous data from the package environment 
}
