% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{createBinNames}
\alias{createBinNames}
\title{A function for creating bin titles.}
\usage{
createBinNames()
}
\value{
{None}
}
\description{
This function creates titles for the bins to represent their ppm range. This function is
meant only for use within the mrbin function.
}
\examples{
\donttest{ createBinNames() }
}
