% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{createBinNumbers}
\alias{createBinNumbers}
\title{A function for creating bin numbers.}
\usage{
createBinNumbers()
}
\value{
{None}
}
\description{
This function calculates numbers of bins from the chosen parameters. This function is
meant only for use within the mrbin function.
}
\examples{
\donttest{ createBinNumbers() }
}
