% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{createBinRegions}
\alias{createBinRegions}
\title{A function for creating bin regions.}
\usage{
createBinRegions()
}
\value{
{None}
}
\description{
This function creates regions for the bins. This function is
meant only for use within the mrbin function.
}
\examples{
\donttest{ createBinRegions() }
}
