% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{createmrbin}
\alias{createmrbin}
\title{A function for creating mrbin objects.}
\usage{
createmrbin()
}
\value{
An (invisible) mrbin object
}
\description{
This function creates an mrbin object and returns it.
}
\examples{
mrbinObject<-createmrbin()
}
