% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{cropNMR}
\alias{cropNMR}
\title{A function for cropping HSQC spectra.}
\usage{
cropNMR(plot = FALSE)
}
\arguments{
\item{plot}{Should a plot of the bins before and after cropping be shown? Defaults to FALSE.}
}
\value{
{None}
}
\description{
This function crops HSQC spectra to the region along the diagonal to remove
uninformative signals. Will work only for 1H-13C HSQC spectra.
}
\examples{
## Requires data
# cropNMR()
}
