% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{.onLoad}
\alias{.onLoad}
\title{A function executed when loading this package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\value{
{None}
}
\description{
This function resets the parameter variables.
}
\examples{
\donttest{ .onLoad() }
}
