% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{down}
\alias{down}
\title{A function for changing plotNMR plots.}
\usage{
down()
}
\value{
{None}
}
\description{
This function moves down the plot region of the current NMR plot (only 2D).
}
\examples{
setParam(parameters=list(dimension="2D",binwidth2D=1,binheight=10,
         NMRfolders=system.file("extdata/1/12/pdata/10",package="mrbin")))
binMultiNMR()
plotNMR()
zoomIn()
down()
}
