% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{getEnv}
\alias{getEnv}
\title{A function for saving the package environment.}
\usage{
getEnv()
}
\value{
A list containing all objects from the local package environment.
}
\description{
This function returns a list of all objects of the current package environment.
This may be helpful for debugging or for accessing NMR spectral data and the raw bin data.
}
\examples{
tempList<-getEnv()
}
