% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{intMin}
\alias{intMin}
\title{A function for changing plotNMR plots.}
\usage{
intMin()
}
\value{
{None}
}
\description{
This function decreases the intensity of the current NMR spectrum plot.
}
\examples{
setParam(parameters=list(dimension="1D",binwidth1D=.1,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
binMultiNMR()
intMin()
}
