% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{mrbin}
\alias{mrbin}
\title{A function setting the parameters and performing binning and data processing}
\usage{
mrbin(silent = FALSE, setDefault = FALSE, parameters = NULL)
}
\arguments{
\item{silent}{If TRUE, the user will be asked no questions and binning and data analysis will run according to the current parameters. Defaults to FALSE.}

\item{setDefault}{If TRUE, all current parameters will be replaced by the default parameters (before loading any provided parameters sets). Defaults to FALSE.}

\item{parameters}{Optional: A list of parameters to be used. If omitted, the user will be asked through a series of question to set the parameters.}
}
\value{
An invisible list containing bins (data after processing), parameters, and factors
}
\description{
This function guides the user through the set-up of parameters, starts binning
and performs the chosen data processing steps
If a list of parameters is provided and silent is set to TRUE, no user input
is requested and binning and data processing are performed silently.
}
\examples{
## Let the user set parameters interactively
# results <- mrbin()
## Set all parameters in the console and run without user intervention
# results <- mrbin(silent=TRUE,
#              setDefault=FALSE,
#              parameters=list(
#              dimension="1D",
#              binMethod="Rectangular bins",
#              binwidth1D=.01,
#              referenceScaling="Yes",
#              removeSolvent="Yes",
#              removeAreas="No",
#              sumBins="No",
#              noiseRemoval="Yes",
#              PQNScaling="Yes",
#              fixNegatives="Yes",
#              logTrafo="Yes",
#              saveFiles="Yes",
#              outputFileName="mrbin_test_results",
#              NMRfolders=c("C:/NMR/Sample01/1/pdata/1",
#                           "C:/NMR/Sample19/1/pdata/1",
#                           "C:/NMR/Sample61/1/pdata/1")
#      ))
## Test example, uses variables instead of files for speed
results <- mrbin(silent=TRUE,
                 parameters=list(dimension="1D",
                        binwidth1D=0.05,
                        signal_to_noise1D=35,
                        PCAtitlelength=20,
                        readFromFile = FALSE,#Only for testing, should be TRUE in all other cases
                        NMRfolders = c("spectrum1",
                                       "spectrum2",
                                       "spectrum3")#Usually this should be folder names
                        )
        )
}
