% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{mrbinGUI}
\alias{mrbinGUI}
\title{A function setting the parameters}
\usage{
mrbinGUI()
}
\description{
This function guides the user through the set-up of parameters and starts mrbin().
}
\examples{
mrbinGUI()
}
