% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{plotBins}
\alias{plotBins}
\title{A function for plotting the current bin positions}
\usage{
plotBins(showtitle = TRUE)
}
\arguments{
\item{showtitle}{If TRUE, the spectrum title is shown. Defaults to TRUE.}
}
\value{
{None}
}
\description{
This function plots the current binned spectrum.
}
\examples{
## Requires data
# plotBins()
}
