% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{plotNMR}
\alias{plotNMR}
\title{A function for plotting NMR spectra.}
\usage{
plotNMR()
}
\description{
This function plots the current NMR spectrum. To change the plot, use zoom(),
zoomIn(), zoomOut(), intPlus(), intMin(), left(), right().
For 2D data use additionally: contMin(), contPlus(), up(), down()
}
\examples{
\donttest{
plotNMR()
}
}
