% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{PQNScaling}
\alias{PQNScaling}
\title{A function for PQN scaling.}
\usage{
PQNScaling()
}
\description{
This function performs PQN scaling. To further exlude unreliable noise, only
the most intense signals are used. For 1H-13C HSQC spectra, most of the sugar
regions are excluded to avoid a dominating ffect of the multiple sugar
signals.
}
\examples{
\donttest{
PQNScaling()
}
}
