% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{printParameters}
\alias{printParameters}
\title{A function for printing parameters to the screen.}
\usage{
printParameters()
}
\value{
{None}
}
\description{
This function reads parameters from the global variable mrbin.env$mrbin$parameters and
prints the required R code for creating a data set to the screen.
}
\examples{
printParameters()
}
