% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{putToEnv}
\alias{putToEnv}
\title{A function for changing and adding variables in the package environment.}
\usage{
putToEnv(variableList)
}
\arguments{
\item{variableList}{A list containing all objects to be saved in the local package environment.}
}
\value{
{None}
}
\description{
This function can change variables in the current package environment.
This may be helpful for debugging or for some plotting functions.
}
\examples{
\donttest{ putToEnv(list(bins=NULL)) }
}
