% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{readBruker}
\alias{readBruker}
\title{A function for reading Bruker NMR spectra.}
\usage{
readBruker()
}
\value{
{None}
}
\description{
This function reads Bruker NMR data.
}
\examples{
## Requires folder names
# readBruker()
}
