% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{readNMR}
\alias{readNMR}
\title{A function for reading NMR spectra.}
\usage{
readNMR()
}
\description{
This function picks the correct NMR reading function, based on vendor.
}
\examples{
## Requires data
# readNMR()
}
