% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{recreatemrbin}
\alias{recreatemrbin}
\title{A function recreating parameters from previous runs.}
\usage{
recreatemrbin(filename = NULL, graphics = TRUE)
}
\arguments{
\item{filename}{File path/name of the mrbin parameter file to be loaded}

\item{graphics}{Controls whether pop-up windows are shown for selections. Defaults to TRUE.}
}
\value{
{None}
}
\description{
This function reads parameters from a text file that was created during a
previous run or mrbin(). After reading, the data can be recreated using
mrbin(). File names in $parameters might need to be updated.
}
\examples{
# Insert full folder path and file name
recreatemrbin(system.file("extdata/mrbin.txt",package="mrbin"))
}
