% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{referenceScaling}
\alias{referenceScaling}
\title{A function for scaling to the reference area.}
\usage{
referenceScaling()
}
\value{
{None}
}
\description{
This function scales NMR data to the reference area.
}
\examples{
resetEnv()
setParam(parameters=list(dimension="1D",referenceScaling="No",
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
binMultiNMR()
referenceScaling()
}
