% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{removeFromPlot}
\alias{removeFromPlot}
\title{A function for removing NMR spectra from the plot list.}
\usage{
removeFromPlot(folder = NULL, dimension = "1D")
}
\arguments{
\item{folder}{Defines the exact NMR data folder.}

\item{dimension}{Defines the data dimension, "1D" or "2D".}
}
\value{
{none}
}
\description{
This function removes a spectrum from the plot list.
}
\examples{
\donttest{ removeFromPlot() }
}
