% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{removeNoise}
\alias{removeNoise}
\title{A function for removing bins below noise level.}
\usage{
removeNoise()
}
\description{
This function checks for each bin (column) whether its level is below the
individual noise level times the signal-to-noise ratio. If less than the
defined threshold level are above noise*SNR, the whole bin is removed.
}
\examples{
removeNoise()
}
