% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{removeSpectrum}
\alias{removeSpectrum}
\title{A function for removing a spectrum.}
\usage{
removeSpectrum()
}
\description{
This function lets the user pick spectra from a list for removal from data
analysis.
}
\examples{
## Will work only after creating a spectrum list
# removeSpectrum()
}
