% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{right}
\alias{right}
\title{A function for changing plotNMR plots.}
\usage{
right()
}
\description{
This function moves right the plot region of the current NMR plot.
}
\examples{
## Requires data
# right()
}
