% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{selectBrukerFolders}
\alias{selectBrukerFolders}
\title{A function for selecting Bruker NMR data folders.}
\usage{
selectBrukerFolders()
}
\description{
This function lets the user set NMR data folders (for Bruker data).
}
\examples{
\donttest{
selectBrukerFolders()
}
}
