% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setFactors}
\alias{setFactors}
\title{A function for setting group members.}
\usage{
setFactors()
}
\description{
This function lets the user pick samples from a list to assign them to
groups.
}
\examples{
## Will work only after setting a sectrum list
# setFactors()
}
