% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setOffset}
\alias{setOffset}
\title{A function for changing plotNMR plots.}
\usage{
setOffset(offsetValue = NULL)
}
\arguments{
\item{offsetValue}{The new offset value. Defaults to NULL}
}
\value{
{None}
}
\description{
This function moves up or down the 1D plot region of the current NMR plot.
}
\examples{
setOffset(0)
}
