% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setParam}
\alias{setParam}
\title{A function setting parameters and checking for consistency.}
\usage{
setParam(parameters = NULL, metadata = NULL)
}
\arguments{
\item{parameters}{List of parameters to be set}

\item{metadata}{List of metadata to be set}
}
\value{
{None}
}
\description{
This function set parameters and checks parameters for consistency.
}
\examples{
setParam(parameters=list(dimension="1D"))
}
