% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectrum1}
\alias{spectrum1}
\title{1H NMR spectral data of a mix of metabolites.}
\format{A vector, names are chemical shift values in ppm}
\source{
\url{http://www.kleinomicslab.com/}
}
\usage{
spectrum1
}
\description{
A dataset containing NMR spectral data of a mix of metabolites
}
\keyword{datasets}
