% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{timeStampMrbin}
\alias{timeStampMrbin}
\title{A function for time stamping mrbin objects.}
\usage{
timeStampMrbin(
  mrbinObject,
  functionName = "InProgress...",
  versionNumber = "0",
  changeDetails = "InProgress...",
  steps = 0,
  comment = ""
)
}
\arguments{
\item{mrbinObject}{An mrbin object}

\item{functionName}{Name of the package and function calling this command}

\item{versionNumber}{Version number of the package calling this command}

\item{changeDetails}{Details of changes made to the mrbin object}

\item{steps}{Indicates which step to perform: 0 (only pre-change), 1 (only post-change)}

\item{comment}{An optional character vector describing the change}
}
\value{
An (invisible) mrbin object
}
\description{
This function adds time stamps to an mrbin object and returns it. Is used only within functions making changes to mrbin objects.
}
\examples{
mrbinObject<-createmrbin()
mrbinObject<-timeStampMrbin(mrbinObject)
}
