% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoomIn}
\alias{zoomIn}
\title{A function for changing plotNMR plots.}
\usage{
zoomIn()
}
\value{
{None}
}
\description{
This function zooms into the plot region of the current NMR plot.
}
\examples{
setParam(parameters=list(dimension="1D",binwidth1D=.1,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
binMultiNMR()
plotNMR()
zoomIn()
}
