% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{binMultiNMR}
\alias{binMultiNMR}
\title{A function for binning multiple NMR spectra.}
\usage{
binMultiNMR()
}
\description{
This function creates bins for each spectrum in mrbinparam$NMRfolders and
saves the bins to mrbinbins.
}
\examples{
binMultiNMR()
}
