% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{binMultiNMR}
\alias{binMultiNMR}
\title{A function for binning multiple NMR spectra.}
\usage{
binMultiNMR()
}
\value{
{None}
}
\description{
This function creates bins for each spectrum in mrbin.env$mrbinparam$NMRfolders and
saves the bins to mrbin.env$bins. Should only be run within the mrbin function.
}
\examples{
\donttest{ binMultiNMR() }
}
