% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{calculateNoise}
\alias{calculateNoise}
\title{A function for calculating noise levels.}
\usage{
calculateNoise()
}
\value{
{None}
}
\description{
This function calculates noise levels.
}
\examples{
mrbinExample<-mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",
                    binwidth1D=0.05,noiseRemoval="No",
                    NMRfolders=c(system.file("extdata/1/10/pdata/10",package="mrbin"),
                                system.file("extdata/2/10/pdata/10",package="mrbin"))))
calculateNoise()
}
