% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{calculateNoise}
\alias{calculateNoise}
\title{A function for calculating noise levels.}
\usage{
calculateNoise()
}
\value{
{None}
}
\description{
This function calculates noise levels for the current spectrum. Only for use
within the mrbin function.
}
\examples{
\donttest{ calculateNoise() }
}
