% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{contMin}
\alias{contMin}
\title{A function for changing plotNMR plots.}
\usage{
contMin()
}
\description{
This function decreases the minimum contour level of the current 2D NMR
spectrum plot.
}
\examples{
## Requires data
# contMin()
}
