% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{contPlus}
\alias{contPlus}
\title{A function for changing plotNMR plots.}
\usage{
contPlus()
}
\description{
This function increases the minimum contour level of the current 2D NMR
spectrum plot.
}
\examples{
\donttest{
contPlus()
}
}
