% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{contPlus}
\alias{contPlus}
\title{A function for changing plotNMR plots.}
\usage{
contPlus(refreshPlot = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}
}
\value{
{None}
}
\description{
This function increases the minimum contour level of the current 2D NMR
spectrum plot.
}
\examples{
resetEnv()
addToPlot(folder=system.file("extdata/1/12/pdata/10",package="mrbin"),dimension="2D")
plotNMR()
contPlus()
}
