% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{cropNMR}
\alias{cropNMR}
\title{A function for cropping HSQC spectra.}
\usage{
cropNMR()
}
\value{
{None}
}
\description{
This function crops HSQC spectra to the region along the diagonal to remove
uninformative signals. Will work only for 1H-13C HSQC spectra.
}
\examples{
resetEnv()
results<-mrbin(silent=TRUE,
         parameters=list(dimension="2D",binwidth2D=1,binheight=4,cropHSQC="No",PCA="No",
         PQNScaling="No",noiseRemoval="No",removeSolvent="No",verbose=TRUE,tryParallel=TRUE,
         NMRfolders=c(system.file("extdata/1/12/pdata/10",package="mrbin"))))
cropNMR()
}
