% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{down}
\alias{down}
\title{A function for changing plotNMR plots.}
\usage{
down()
}
\description{
This function moves down the plot region of the current NMR plot (only 2D).
}
\examples{
## Requires data
# down()
}
