% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{down}
\alias{down}
\title{A function for changing plotNMR plots.}
\usage{
down(refreshPlot = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}
}
\value{
{None}
}
\description{
This function moves down the plot region of the current NMR plot (only 2D).
}
\examples{
resetEnv()
mrbin(silent=TRUE,parameters=list(dimension="2D",binwidth2D=0.5,
         binheight=3,PQNScaling="No",referenceScaling="No",binRegion=c(4,3,60,65),
         noiseRemoval="No",trimZeros="No",cropHSQC="No",tryParallel=TRUE,
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,saveFiles="No",
         NMRfolders=c(system.file("extdata/1/12/pdata/10",package="mrbin"))))
plotNMR()
zoomIn()
down()
}
