% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{intMin}
\alias{intMin}
\title{A function for changing plotNMR plots.}
\usage{
intMin()
}
\description{
This function decreases the intensity of the current NMR spectrum plot.
}
\examples{
## Requires data
# intMin()
}
