% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{intPlus}
\alias{intPlus}
\title{A function for changing plotNMR plots.}
\usage{
intPlus()
}
\value{
{None}
}
\description{
This function increases the intensity of the current NMR spectrum plot.
}
\examples{
## Requires data
# intPlus()
}
