% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{left}
\alias{left}
\title{A function for changing plotNMR plots.}
\usage{
left(refreshPlot = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}
}
\value{
{None}
}
\description{
This function moves left the plot region of the current NMR plot.
}
\examples{
resetEnv()#clean up previous data from the package environment 
mrbin(silent=TRUE,parameters=list(dimension="1D",binwidth1D=.5,
         noiseRemoval="No",trimZeros="No",
         PQNScaling="No",saveFiles="No",referenceScaling="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,
         NMRvendor="mrbin",
         example=TRUE,#only used for the package examples
         NMRfolders=system.file("extdata/1.mr1",package="mrbin")))
plotNMR()
zoomIn()
left()
}
