% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{left}
\alias{left}
\title{A function for changing plotNMR plots.}
\usage{
left(refreshPlot = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}
}
\value{
{None}
}
\description{
This function moves left the plot region of the current NMR plot.
}
\examples{
resetEnv()
mrbin(silent=TRUE,parameters=list(dimension="1D",binwidth1D=.5,
         noiseRemoval="No",trimZeros="No",tryParallel=TRUE,
         PQNScaling="No",saveFiles="No",referenceScaling="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoomIn()
left()
}
