% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{logTrafo}
\alias{logTrafo}
\title{A function for log transforming data.}
\usage{
logTrafo()
}
\value{
{None}
}
\description{
This function simply log transforms. Will not work with negative data.
}
\examples{
## Will work only after creating bins
# logTrafo()
}
