% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{logTrafo}
\alias{logTrafo}
\title{A function for log transforming data.}
\usage{
logTrafo()
}
\value{
{None}
}
\description{
This function simply log transforms. Will not work with negative data.
}
\examples{
mrbinExample<-mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D", logTrafo="No",
                    binwidth1D=0.05,signal_to_noise1D=50, verbose=TRUE, PCA="No",
                    NMRfolders=c(system.file("extdata/1/10/pdata/10",package="mrbin"),
                                system.file("extdata/2/10/pdata/10",package="mrbin"))))
logTrafo()
}
