% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{metadatamrbin}
\alias{metadatamrbin}
\title{A function for interactively editing metadata of mrbin objects.}
\usage{
metadatamrbin(mrbinResults, metadata = NULL, graphics = graphics)
}
\arguments{
\item{mrbinResults}{An mrbin object}

\item{metadata}{An optional list of objects to be changed. If provided, interactive mode is deactivated}

\item{graphics}{Controls whether pop-up windows are shown for selections. Defaults to TRUE.}
}
\value{
An invisible mrbin object
}
\description{
This function edits interactively or non-interactively the metadata filed of the provided mrbin object.
}
\examples{
 results<-mrbin(silent=TRUE,
                   parameters=list(verbose=TRUE,dimension="1D",PQNScaling="No",
                   binwidth1D=0.04,signal_to_noise1D=1,PCA="No",binRegion=c(9.5,0.5,10,156),
                   saveFiles="No",referenceScaling="No",noiseRemoval="No",
                   fixNegatives="No",logTrafo="No",noiseThreshold=.05,tryParallel=TRUE,
                   NMRfolders=c(system.file("extdata/2/10/pdata/10",package="mrbin"),
                              system.file("extdata/3/10/pdata/10",package="mrbin"))
                   ))
 results<-metadatamrbin(results,metadata=list(projectTitle="Test project"))
}
