% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{mrbin}
\alias{mrbin}
\title{A function setting the parameters}
\usage{
mrbin()
}
\value{
An invisible list containing $bins (data after processing), $parameters
}
\description{
This function guides the user through the set-up of parameters and starts mrbin().
}
