% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{plotBins}
\alias{plotBins}
\title{A function for plotting the current 2D bin data.}
\usage{
plotBins()
}
\description{
This function plots the current binned spectrum. Works only for 2D data.
}
\examples{
## Requires data
# plotBins()
}
