% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{plotBins}
\alias{plotBins}
\title{A function for plotting the current bin positions}
\usage{
plotBins(showtitle = TRUE)
}
\arguments{
\item{showtitle}{If TRUE, the spectrum title is shown. Defaults to TRUE.}
}
\value{
{None}
}
\description{
This function plots the current binned spectrum.
}
\examples{
mrbinExample<-mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",
                    binwidth1D=0.05,PQNScaling="No",PCA="No",
                    NMRfolders=c(system.file("extdata/1/10/pdata/10",package="mrbin"))))
plotBins()
}
