% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{plotResults}
\alias{plotResults}
\title{A function for plotting quality indicators, including PCA plots.}
\usage{
plotResults()
}
\description{
This function plots boxplots (bin-wise and sample-wise) as visual quality indicators. It also performs PCA, then plots PC1 and PC2 and loading plots.
}
\examples{
\donttest{
plotResults()
}
}
