% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{PQNScaling}
\alias{PQNScaling}
\title{A function for PQN scaling.}
\usage{
PQNScaling()
}
\value{
{None}
}
\description{
This function performs PQN scaling. To further exclude unreliable noise, only
the most intense signals are used. For 1H-13C HSQC spectra, most of the sugar
regions are excluded to avoid a dominating effect of the multiple sugar
signals.
}
\examples{
mrbinExample<-mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",
                    binwidth1D=0.05,PQNScaling="No",PCA="No",
                    NMRfolders=c(system.file("extdata/1/10/pdata/10",package="mrbin"),
                                system.file("extdata/2/10/pdata/10",package="mrbin"),
                                system.file("extdata/3/10/pdata/10",package="mrbin"))))
PQNScaling()
}
