% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{readNMR}
\alias{readNMR}
\title{A function for reading NMR spectra.}
\usage{
readNMR()
}
\value{
{None}
}
\description{
This function picks the correct NMR reading function, based on vendor. This function is
meant only for use within the mrbin function.
}
\examples{
\donttest{ readNMR() }
}
