% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{referenceScaling}
\alias{referenceScaling}
\title{A function for scaling to the reference area.}
\usage{
referenceScaling()
}
\description{
This function scales NMR data to the reference area.
}
\examples{
\donttest{
referenceScaling()
}
}
