% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{removeAreas}
\alias{removeAreas}
\title{A function for removing additional regions.}
\usage{
removeAreas()
}
\description{
This function removes additional regions. This can be useful when some areas
are visibly affected by spectral artifacts.
}
\examples{
## Requires data
# removeAreas()
}
