% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{removeSolvent}
\alias{removeSolvent}
\title{A function for removing the solvent region.}
\usage{
removeSolvent()
}
\value{
{None}
}
\description{
This function removes the solvent region. Should only be run within the mrbin function.
}
\examples{
\donttest{ removeSolvent() }
}
