% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{right}
\alias{right}
\title{A function for changing plotNMR plots.}
\usage{
right()
}
\value{
{None}
}
\description{
This function moves right the plot region of the current NMR plot.
}
\examples{
mrbin(silent=TRUE,parameters=list(dimension="1D",binwidth1D=.5,
         PQNScaling="No",saveFiles="No",referenceScaling="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=FALSE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoomIn()
right()
}
