% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{selectBrukerFolders}
\alias{selectBrukerFolders}
\title{A function for selecting Bruker NMR data folders.}
\usage{
selectBrukerFolders()
}
\value{
An invisible list of folder names, or "Go back" or "stop"
}
\description{
This function lets the user set NMR data folders interactively (for Bruker data). This function
is meant only for use within the mrbin function.
}
\examples{
\donttest{ selectBrukerFolders() }
}
