% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{selectFolders}
\alias{selectFolders}
\title{A function for selecting NMR data folders.}
\usage{
selectFolders()
}
\value{
{None}
}
\description{
This function selects the correct folder selection function for the vendor
(currently only Bruker).
}
\examples{
## Requires user input
# selectFolders()
}
