% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{selectFolders}
\alias{selectFolders}
\title{A function for selecting NMR data folders.}
\usage{
selectFolders()
}
\value{
An invisible list of folder names, or "Go back" or "stop"
}
\description{
This function selects the correct folder selection function for the vendor
(currently only Bruker). This function is meant only for use within the
mrbin function.
}
\examples{
\donttest{ selectFolders() }
}
