% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setCurrentSpectrum}
\alias{setCurrentSpectrum}
\title{A function for setting the current spectrum.}
\usage{
setCurrentSpectrum()
}
\description{
This function lets the user pick a spectrum from the list of spectra
analysis.
}
\examples{
\donttest{
setCurrentSpectrum()
}
}
