% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setCurrentSpectrum}
\alias{setCurrentSpectrum}
\title{A function for interactively setting the current spectrum.}
\usage{
setCurrentSpectrum()
}
\value{
{None}
}
\description{
This function lets the user pick a spectrum from the list of spectra
analysis. This function is meant only for use within the mrbin function.
}
\examples{
\donttest{ setCurrentSpectrum() }
}
