% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{setFactors}
\alias{setFactors}
\title{A function for setting group members.}
\usage{
setFactors()
}
\value{
{None}
}
\description{
This function lets the user pick samples from a list to assign them to
groups. This function is meant only for use within the mrbin function.
}
\examples{
\donttest{ setFactors() }
}
