% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{up}
\alias{up}
\title{A function for changing plotNMR plots.}
\usage{
up()
}
\value{
{None}
}
\description{
This function moves up the plot region of the current NMR plot (only 2D).
}
\examples{
## Requires data
# up()
}
