% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{up}
\alias{up}
\title{A function for changing plotNMR plots.}
\usage{
up(refreshPlot = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}
}
\value{
{None}
}
\description{
This function moves up the plot region of the current NMR plot (only 2D).
}
\examples{
resetEnv()#clean up previous data from the package environment 
mrbin(silent=TRUE,parameters=list(dimension="2D",binwidth2D=0.5,
         binheight=3,PQNScaling="No",referenceScaling="No",binRegion=c(4,3,60,65),
         noiseRemoval="No",trimZeros="No",cropHSQC="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,saveFiles="No",
         example=TRUE,#only used for the package examples
         NMRfolders=system.file("extdata/1.mr2",package="mrbin"),
         NMRvendor="mrbin"))
plotNMR()
zoomIn()
up()
}
