% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{up}
\alias{up}
\title{A function for changing plotNMR plots.}
\usage{
up()
}
\value{
{None}
}
\description{
This function moves up the plot region of the current NMR plot (only 2D).
}
\examples{
mrbin(silent=TRUE,parameters=list(dimension="2D",binwidth2D=0.5,
         binheight=20,PQNScaling="No",referenceScaling="No",binRegion=c(8,2,20,140),
         fixNegatives="No",logTrafo="No",PCA="No",verbose=FALSE,saveFiles="No",
         NMRfolders=c(system.file("extdata/1/12/pdata/10",package="mrbin"))))
plotNMR()
zoomIn()
up()
}
