% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoom}
\alias{zoom}
\title{A function for changing plotNMR plots.}
\usage{
zoom(left = NULL, right = NULL, top = NULL, bottom = NULL)
}
\arguments{
\item{left}{New left boundary}

\item{right}{New right boundary}

\item{top}{New top boundary}

\item{bottom}{New bottom boundary}
}
\value{
{None}
}
\description{
This function changes the plot region of the current NMR plot.
}
\examples{
# Ask for user input
\donttest{ zoom() }
mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",binwidth1D=.1,
         PQNScaling="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=FALSE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoom(left=4.6,right=2,top=10,bottom=150)
}
