% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoom}
\alias{zoom}
\title{A function for changing plotNMR plots.}
\usage{
zoom(
  left = NULL,
  right = NULL,
  top = NULL,
  bottom = NULL,
  refreshPlot = TRUE,
  dimension = "2D"
)
}
\arguments{
\item{left}{New left boundary}

\item{right}{New right boundary}

\item{top}{New top boundary}

\item{bottom}{New bottom boundary}

\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}

\item{dimension}{Dimension of the data. Defaults to "2D"}
}
\value{
An invisible value indicating if a change occurred
}
\description{
This function changes the plot region of the current NMR plot. Can be called with
no arguments: zoom(). In this case the user will be asked for manual input.
}
\examples{
mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",binwidth1D=.1,
         PQNScaling="No",noiseRemoval="No",trimZeros="No",tryParallel=TRUE,
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoom(left=4.6,right=2,top=10,bottom=150)
}
