% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoomIn}
\alias{zoomIn}
\title{A function for changing plotNMR plots.}
\usage{
zoomIn()
}
\value{
{None}
}
\description{
This function zooms into the plot region of the current NMR plot.
}
\examples{
## Requires data
# zoomIn()
}
