% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoomIn}
\alias{zoomIn}
\title{A function for changing plotNMR plots.}
\usage{
zoomIn(refreshPlot = TRUE, x = TRUE, y = TRUE)
}
\arguments{
\item{refreshPlot}{Refresh plot automatically. Defaults to TRUE}

\item{x}{Change x axis? Defaults to TRUE}

\item{y}{Change y axis? Defaults to TRUE}
}
\value{
{None}
}
\description{
This function zooms into the plot region of the current NMR plot.
}
\examples{
mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",binwidth1D=.1,
         PQNScaling="No",noiseRemoval="No",trimZeros="No",tryParallel=TRUE,
         fixNegatives="No",logTrafo="No",PCA="No",verbose=TRUE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoomIn()
}
