% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoomOut}
\alias{zoomOut}
\title{A function for changing plotNMR plots.}
\usage{
zoomOut()
}
\description{
This function zooms out from the plot region of the current NMR plot.
}
\examples{
## Requires data
# zoomOut()
}
