% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbin.R
\name{zoomOut}
\alias{zoomOut}
\title{A function for changing plotNMR plots.}
\usage{
zoomOut()
}
\value{
{None}
}
\description{
This function zooms out from the plot region of the current NMR plot.
}
\examples{
mrbin(silent=TRUE,setDefault=TRUE,parameters=list(dimension="1D",binwidth1D=.1,
         PQNScaling="No",
         fixNegatives="No",logTrafo="No",PCA="No",verbose=FALSE,
         NMRfolders=system.file("extdata/1/10/pdata/10",package="mrbin")))
plotNMR()
zoomIn()
zoomOut()
}
